Tried implementing the Goal Decrement
I tied to implement the goal decrement when a player hits the puck more than once before the other player does however I want not able to do this. I am not sure if this was a UI issue where it would not update on the game screen but I believe the code implementation is correct.